requirements.txt
attrs==19.3.0
aws-cdk.assets==1.54.0
aws-cdk.aws-apigateway==1.54.0
aws-cdk.aws-applicationautoscaling==1.54.0
aws-cdk.aws-autoscaling==1.54.0
aws-cdk.aws-autoscaling-common==1.54.0
aws-cdk.aws-autoscaling-hooktargets==1.54.0
aws-cdk.aws-batch==1.54.0
aws-cdk.aws-certificatemanager==1.54.0
aws-cdk.aws-cloudformation==1.54.0
aws-cdk.aws-cloudfront==1.54.0
aws-cdk.aws-cloudwatch==1.54.0
aws-cdk.aws-codeguruprofiler==1.54.0
aws-cdk.aws-cognito==1.54.0
aws-cdk.aws-dynamodb==1.54.0
aws-cdk.aws-ec2==1.54.0
aws-cdk.aws-ecr==1.54.0
aws-cdk.aws-ecr-assets==1.54.0
aws-cdk.aws-ecs==1.54.0
aws-cdk.aws-efs==1.54.0
aws-cdk.aws-elasticloadbalancing==1.54.0
aws-cdk.aws-elasticloadbalancingv2==1.54.0
aws-cdk.aws-events==1.54.0
aws-cdk.aws-glue==1.54.0
aws-cdk.aws-iam==1.54.0
aws-cdk.aws-kms==1.54.0
aws-cdk.aws-lambda==1.54.0
aws-cdk.aws-logs==1.54.0
aws-cdk.aws-route53==1.54.0
aws-cdk.aws-route53-targets==1.54.0
aws-cdk.aws-s3==1.54.0
aws-cdk.aws-s3-assets==1.54.0
aws-cdk.aws-s3-notifications==1.54.0
aws-cdk.aws-sam==1.54.0
aws-cdk.aws-secretsmanager==1.54.0
aws-cdk.aws-servicediscovery==1.54.0
aws-cdk.aws-sns==1.54.0
aws-cdk.aws-sns-subscriptions==1.54.0
aws-cdk.aws-sqs==1.54.0
aws-cdk.aws-ssm==1.54.0
aws-cdk.aws-stepfunctions==1.54.0
aws-cdk.aws-stepfunctions-tasks==1.54.0
aws-cdk.cloud-assembly-schema==1.54.0
aws-cdk.core==1.54.0
aws-cdk.custom-resources==1.54.0
aws-cdk.cx-api==1.54.0
aws-cdk.region-info==1.54.0
awscli==1.18.97
boto3==1.14.20
boto3-type-annotations==0.3.1
botocore==1.17.20
cattrs==1.0.0
colorama==0.4.3
constructs==3.0.4
docutils==0.15.2
jmespath==0.10.0
jsii==1.9.0
numpy==1.19.0
publication==0.0.3
pyasn1==0.4.8
python-dateutil==2.8.1
PyYAML==5.3.1
rsa==4.5
s3transfer==0.3.3
six==1.15.0
typing-extensions==3.7.4.2
urllib3==1.25.9
